refactor: 함수를 모듈로 분리
