+++ src/display/index.js
@@ -1,0 +1,2 @@
+export { default as displayGameResult } from "./displayGameResult.js";
+export { default as displayHint } from "./displayHint.js";
