+++ src/utils/getRandomNumber.js
@@ -1,0 +1,7 @@
+const getRandomNumber = (startNum, endNum) => {
+  const randomNumber = Math.floor(Math.random() * endNum) + startNum;
+
+  return randomNumber;
+};
+
+export default getRandomNumber;
