+++ src/utils/index.js
@@ -1,0 +1,3 @@
+export { default as getRandomNumber } from "./getRandomNumber.js";
+export { default as validateNumber } from "./validateNumber.js";
+export { default as validateRange } from "./validateRange.js";
